clean up/normalize file names
