--- entry/devscan.go
@@ -1,7 +1,0 @@
-package entry
-
-import "zikichombo.org/sio"
-
-type DevScanner interface {
-	DevScan() []*sio.Dev
-}
